requirements.txt
###################################
# Required packages for PyQchem   #
###################################

numpy>=1.8.2
lxml
requests
scipy
matplotlib
PyYAML
six